+++ src/components/Favorites/Favorites.js
