+++ src/components/main.js
@@ -1,8 +1,9 @@
 import React from 'react';
-import LandingPage from './Landingpage';
-import News from './news';
-import Favorites from './Favorites'
-import Login from './Login'
+import LandingPage from './Landingpage/Landingpage';
+import News from './News/news';
+import Favorites from './Favorites/Favorites'
+import Login from './Login/Login'
+import Signup from './Signup/Signup'
 import { Switch, Route } from 'react-router-dom'
 
 const Main = () => (
@@ -11,6 +12,7 @@ const Main = () => (
         <Route path="/news" component={News} />
         <Route path="/favorites" component={Favorites} />
         <Route path="/login" component={Login}/>
+        <Route path="/signup" component={Signup} />
     </Switch>
 )
